implementação do funcionario, responsavel departamento e e departamento.

+++ src/main/java/br/com/rodrigo/api/service/DepartamentoService.java
@@ -1,22 +1,20 @@
 package br.com.rodrigo.api.service;
 
-import br.com.rodrigo.api.exception.ObjetoNaoEncontradoException;
 import br.com.rodrigo.api.exception.ViolocaoIntegridadeDadosException;
 import br.com.rodrigo.api.model.Departamento;
-import br.com.rodrigo.api.model.Funcionario;
-import br.com.rodrigo.api.model.ResponsavelDepartamento;
 import br.com.rodrigo.api.model.dto.DepartamentoDto;
+import br.com.rodrigo.api.repository.CargoRepository;
 import br.com.rodrigo.api.repository.DepartamentoRepository;
-import br.com.rodrigo.api.repository.FuncionarioRepository;
 import br.com.rodrigo.api.repository.ResponsavelDepartamentoRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
+import static br.com.rodrigo.api.exception.ValidationError.ERRO_DELETAR_DERPARTAMENTO_CARGO;
 import static br.com.rodrigo.api.exception.ValidationError.ERRO_DEPARTAMENTO_NAO_ENCONTRADO;
-import static br.com.rodrigo.api.exception.ValidationError.ERRO_FUNCINARIO_NAO_ENCONTRADO;
 
 @Service
 @RequiredArgsConstructor
@@ -24,12 +22,20 @@ public class DepartamentoService {
 
     private final DepartamentoRepository departamentoRepository;
 
-    private final FuncionarioRepository funcionarioRepository;
+    private final CargoRepository cargoRepository;
 
     private final ResponsavelDepartamentoRepository responsavelDepartamentoRepository;
 
     public List<Departamento> listarTodosDepartamentos() {
         return departamentoRepository.findAll();
+    }
+
+    public List<Departamento> listarDepartamentosSemResponsavel() {
+        // Listar apenas departamentos que não possuem responsável
+        return departamentoRepository.findAll()
+                .stream()
+                .filter(departamento -> !responsavelDepartamentoRepository.existsByDepartamentoId(departamento.getId()))
+                .collect(Collectors.toList());
     }
 
     public Optional<Departamento> getDepartamentoById(Long id) {
@@ -46,30 +52,18 @@ public class DepartamentoService {
     public Departamento atualizarDepartamento(Long id, DepartamentoDto departamentoDto) {
         Optional<Departamento> optionalDepartamento = departamentoRepository.findById(id);
 
-        Departamento departamento = optionalDepartamento.map(dep -> {
+        return optionalDepartamento.map(dep -> {
             dep.setNome(departamentoDto.getNome());
             dep.setDescricao(departamentoDto.getDescricao());
             return departamentoRepository.save(dep);
         }).orElseThrow(() -> new ViolocaoIntegridadeDadosException(ERRO_DEPARTAMENTO_NAO_ENCONTRADO + id));
-
-        return departamento;
     }
 
     public void deleteDepartamento(Long id) {
+
+        if (cargoRepository.existsByDepartamentoId(id)) {
+            throw new ViolocaoIntegridadeDadosException(ERRO_DELETAR_DERPARTAMENTO_CARGO);
+        }
         departamentoRepository.deleteById(id);
-    }
-
-    public void vincularResponsavel(Long idDepartamento, Long idFuncionario) {
-        Departamento departamento = departamentoRepository.findById(idDepartamento)
-                .orElseThrow(() -> new ObjetoNaoEncontradoException(ERRO_DEPARTAMENTO_NAO_ENCONTRADO));
-
-        Funcionario funcionario = funcionarioRepository.findById(idFuncionario)
-                .orElseThrow(() -> new ObjetoNaoEncontradoException(ERRO_FUNCINARIO_NAO_ENCONTRADO));
-
-        ResponsavelDepartamento responsavelDepartamento = new ResponsavelDepartamento();
-        responsavelDepartamento.setDepartamento(departamento);
-        responsavelDepartamento.setFuncionario(funcionario);
-
-        responsavelDepartamentoRepository.save(responsavelDepartamento);
     }
 }
